requirements.txt
pytest
requests
selene==2.0.0rc2
allure-pytest==2.9.45
